basic testing of event creation and workflow

+++ yadarts-usb/src/main/java/spare/n52/yadarts/usb/USBEventProducer.java
@@ -63,7 +63,7 @@ public class USBEventProducer implements EventProducer, USBEventReceiver {
 	}
 	
 	@Override
-	public void processEvent(final byte[] rawData) {
+	public void processEvent(final int[] rawData) {
 		executor.submit(new Runnable() {
 			
 			@Override
